Rename 'verify' to 'checksum'
