+++ bag_verifier/src/main/scala/weco/storage_service/bag_verifier/fixity/bag/BagExpectedFixity.scala
@@ -12,7 +12,7 @@ import weco.storage_service.bag_verifier.storage.{Locatable, Resolvable}
 import weco.storage_service.bag_verifier.storage.bag.BagLocatable
 import weco.storage_service.bagit.models._
 import weco.storage_service.bagit.services.BagMatcher
-import weco.storage_service.verify._
+import weco.storage_service.checksum._
 import weco.storage.{Location, Prefix}
 
 class BagExpectedFixity[BagLocation <: Location, BagPrefix <: Prefix[
