+++ bag_verifier/src/test/scala/weco/storage_service/bag_verifier/fixity/FixityCheckerTests.scala
@@ -13,7 +13,7 @@ import weco.storage_service.bag_verifier.storage.{
   LocateFailure,
   LocationParsingError
 }
-import weco.storage_service.verify.{Checksum, ChecksumValue, MD5}
+import weco.storage_service.checksum.{Checksum, ChecksumValue, MD5}
 import weco.storage._
 import weco.storage.generators.{MemoryLocationGenerators, StreamGenerators}
 import weco.storage.providers.memory.{MemoryLocation, MemoryLocationPrefix}
