+++ common/src/main/scala/weco/storage_service/bagit/models/MatchedLocation.scala
@@ -1,6 +1,6 @@
 package weco.storage_service.bagit.models
 
-import weco.storage_service.verify.Checksum
+import weco.storage_service.checksum.Checksum
 
 case class MatchedLocation(
   bagPath: BagPath,
