+++ common/src/main/scala/weco/storage_service/checksum/ChecksumAlgorithm.scala
@@ -1,4 +1,4 @@
-package weco.storage_service.verify
+package weco.storage_service.checksum
 
 import org.apache.commons.codec.digest.MessageDigestAlgorithms
