+++ common/src/main/scala/weco/storage_service/checksum/ChecksumValue.scala
@@ -1,4 +1,4 @@
-package weco.storage_service.verify
+package weco.storage_service.checksum
 
 import weco.json.{TypedString, TypedStringOps}
