+++ common/src/main/scala/weco/storage_service/checksum/Hasher.scala
@@ -1,4 +1,4 @@
-package weco.storage_service.verify
+package weco.storage_service.checksum
 
 import java.io.InputStream
 import java.security.MessageDigest
