+++ common/src/main/scala/weco/storage_service/storage/models/StorageManifest.scala
@@ -4,7 +4,7 @@ import java.time.Instant
 
 import weco.storage_service.bagit.models.{BagId, BagInfo, BagVersion}
 import weco.storage_service.ingests.models.IngestID
-import weco.storage_service.verify.{ChecksumValue, ChecksumAlgorithm}
+import weco.storage_service.checksum.{ChecksumValue, ChecksumAlgorithm}
 
 case class StorageManifestFile(
   checksum: ChecksumValue,
