+++ common/src/test/scala/weco/storage_service/checksum/HasherTest.scala
@@ -1,4 +1,4 @@
-package weco.storage_service.verify
+package weco.storage_service.checksum
 
 import java.io.{FilterInputStream, InputStream}
