+++ indexer/file_indexer/src/test/scala/weco/storage_service/indexer/files/models/IndexedFileTest.scala
@@ -9,7 +9,7 @@ import weco.storage_service.storage.models.{
   PrimaryS3StorageLocation,
   StorageManifestFile
 }
-import weco.storage_service.verify.{SHA256, SHA512}
+import weco.storage_service.checksum.{SHA256, SHA512}
 import weco.storage_service.indexer.models.FileContext
 import weco.storage.s3.{S3ObjectLocation, S3ObjectLocationPrefix}
